+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## src/popup
 
-- add select field on top of rule title and regex check box
+- added yt category select in add rule